+++ buildSrc/src/main/kotlin/BuildModules.kt
@@ -5,7 +5,7 @@ object BuildModules {
     object Features {
         const val ACCOUNT = ":features:account"
         const val HOME = ":features:home"
-        const val LEARN = ":features:learn"
+        const val DISCOVER = ":features:discover"
         const val PROFILE = ":features:profile"
         const val LESSON = ":features:lesson"
     }
