+++ features/discover/.gitignore
