+++ features/discover/src/main/graphql/.graphqlconfig
