+++ features/discover/src/main/kotlin/com/teacherworkout/features/discover/landing/Theme.kt
@@ -1,0 +1,6 @@
+package com.teacherworkout.features.discover.landing
+
+data class Theme(
+    val id: String,
+    val title: String
+)
