+++ features/discover/src/main/kotlin/com/teacherworkout/features/discover/landing/ThemesRepository.kt
@@ -1,0 +1,5 @@
+package com.teacherworkout.features.discover.landing
+
+interface ThemesRepository {
+    suspend fun all(): Result<List<Theme>>
+}
