+++ features/discover/src/main/res/values/strings.xml
