+++ features/home/build.gradle.kts
@@ -10,7 +10,7 @@ android {
     addProductFlavours(this)
 }
 dependencies {
-    implementation(project(BuildModules.Features.LEARN))
+    implementation(project(BuildModules.Features.DISCOVER))
     implementation(project(BuildModules.Features.PROFILE))
     implementation(project(BuildModules.Features.LESSON))
     implementation(project(BuildModules.Commons.UI))
