+++ features/home/src/main/kotlin/com/teacherworkout/features/home/di/KoinModules.kt
@@ -1,10 +1,10 @@
 package com.teacherworkout.features.home.di
 
 import com.teacherworkout.commons.ui.data.impl.FakeLessonsRepository
-import com.teacherworkout.features.home.lessons.HomeViewModel
+import com.teacherworkout.features.home.landing.LandingViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val homeModule = module {
-    viewModel { HomeViewModel(FakeLessonsRepository()) }
+    viewModel { LandingViewModel(FakeLessonsRepository()) }
 }
