--- features/learn/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.teacherworkout.features.learn">
-
-</manifest>
