--- features/learn/src/main/kotlin/com/teacherworkout/features/learn/discover/DiscoveryModule.kt
@@ -1,8 +1,0 @@
-package com.teacherworkout.features.learn.discover
-
-import org.koin.androidx.viewmodel.dsl.viewModel
-import org.koin.dsl.module
-
-val discoveryModule = module {
-    viewModel { DiscoverViewModel(GraphqlThemesRepository(get())) }
-}
