--- features/learn/src/main/kotlin/com/teacherworkout/features/learn/discover/Theme.kt
@@ -1,6 +1,0 @@
-package com.teacherworkout.features.learn.discover
-
-data class Theme(
-    val id: String,
-    val title: String
-)
