--- features/learn/src/main/kotlin/com/teacherworkout/features/learn/discover/ThemesRepository.kt
@@ -1,5 +1,0 @@
-package com.teacherworkout.features.learn.discover
-
-interface ThemesRepository {
-    suspend fun all(): Result<List<Theme>>
-}
