+++ features/profile/src/main/kotlin/com/teacherworkout/features/profile/ProfileFeature.kt
@@ -6,14 +6,15 @@ import androidx.navigation.compose.composable
 import androidx.navigation.navigation
 import com.teacherworkout.commons.ui.composables.BottomBarScaffold
 import com.teacherworkout.commons.ui.navigation.AppDestinations
+import com.teacherworkout.features.profile.landing.Screen
 
 fun NavGraphBuilder.profileFeature(navHostController: NavHostController) {
     navigation(
-        startDestination = AppDestinations.Profile.Profile.route,
+        startDestination = AppDestinations.Profile.Landing.route,
         route = AppDestinations.Features.profile
     ) {
-        composable(route = AppDestinations.Profile.Profile.route) {
-            BottomBarScaffold(navHostController) { ProfileScreen() }
+        composable(route = AppDestinations.Profile.Landing.route) {
+            BottomBarScaffold(navHostController) { Screen() }
         }
     }
 }
